% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetry.R
\name{asymmetry}
\alias{asymmetry}
\title{Asymmetry Function}
\usage{
asymmetry(delta, z, error.check = FALSE, echo = FALSE)
}
\arguments{
\item{delta}{an n by n square asymmetric hollow matrix containing dissimilarities.}

\item{z}{coordinates (n by p) after symmetric analysis}

\item{error.check}{extensive check validity input parameters (default = FALSE).}

\item{echo}{print (intermediate) results (default = FALSE).}
}
\value{
ssaf skew-symmetric sum-of-squares-accounted-for

vaf skew-symmetric variance-accounted-for

drifts drift vectors starting from coordinates in z (n by p matrix)

radii n vector with circle radii
}
\description{
\code{asymmetry} return statistics from asymmetry analyses.
}
\references{
Gower (1968).
}
\author{
Frank M.T.A. Busing
}
