% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilation.R
\name{dilation}
\alias{dilation}
\title{Dilation}
\usage{
dilation(source, weights = NULL, target = NULL, error.check = FALSE)
}
\arguments{
\item{source}{n x m source matrix}

\item{weights}{weights matrix, size n}

\item{target}{if NULL: rotate source to principal axes; otherwise: rotate source to n x m target}

\item{error.check}{extensive check validity input parameters (default = FALSE).}
}
\value{
dilation factor
}
\description{
\code{dilation} returns dilation or scale factor,
such that || factor * source - target ||^2_weights is minimal.
}
\references{
Gower (1968).
            Commandeur (1991)
}
\author{
Frank M.T.A. Busing
}
