% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.fmds.R
\name{explain.fmds}
\alias{explain.fmds}
\title{Explain method for all fmds objects}
\usage{
\method{explain}{fmds}(
  x,
  q,
  w = NULL,
  level = c("absolute", "ratio", "linear", "ordinal", "nominal"),
  MAXITER = 1024,
  FCRIT = 1e-08,
  error.check = FALSE,
  echo = FALSE,
  ...
)
}
\arguments{
\item{x}{fmds object}

\item{q}{variables}

\item{w}{weights}

\item{level}{transformation level}

\item{MAXITER}{maximum number of iterations}

\item{FCRIT}{convergence criterion}

\item{error.check}{error chacking}

\item{echo}{intermediate output}

\item{\ldots}{additional arguments to be passed.}
}
\value{
data data

weights weights

transformed.data transformed data

approach approach

degree zero

ninner zero

iknots NULL

anchor anchor

knotstype zero

coordinates NULL

coefficients coefficients

distances distances

last.iteration last.iteration

last.difference last.difference

n.stress n.stress

stress.1 stress.1

call call
}
\description{
\code{explain} fits variables to coordinate configuration
}
