% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food.R
\docType{data}
\name{food}
\alias{food}
\title{Food data}
\format{
45 x 45 dissimilarity matrix
\itemize{
  \item V1: dissimilarities for V1.
  \item V2: dissimilarities for V2.
  \item V3: dissimilarities for V3.
  \item V4: dissimilarities for V4.
  \item V5: dissimilarities for V5.
  \item V6: dissimilarities for V6.
  \item V7: dissimilarities for V7.
  \item V8: dissimilarities for V8.
  \item V9: dissimilarities for V9.
  \item V10: dissimilarities for V10.
  \item V11: dissimilarities for V11.
  \item V12: dissimilarities for V12.
  \item V13: dissimilarities for V13.
  \item V14: dissimilarities for V14.
  \item V15: dissimilarities for V15.
  \item V16: dissimilarities for V16.
  \item V17: dissimilarities for V17.
  \item V18: dissimilarities for V18.
  \item V19: dissimilarities for V19.
  \item V20: dissimilarities for V20.
  \item V21: dissimilarities for V21.
  \item V22: dissimilarities for V22.
  \item V23: dissimilarities for V23.
  \item V24: dissimilarities for V24.
  \item V25: dissimilarities for V25.
  \item V26: dissimilarities for V26.
  \item V27: dissimilarities for V27.
  \item V28: dissimilarities for V28.
  \item V29: dissimilarities for V29.
  \item V30: dissimilarities for V30.
  \item V31: dissimilarities for V31.
  \item V32: dissimilarities for V32.
  \item V33: dissimilarities for V33.
  \item V34: dissimilarities for V34.
  \item V35: dissimilarities for V35.
  \item V36: dissimilarities for V36.
  \item V37: dissimilarities for V37.
  \item V38: dissimilarities for V38.
  \item V39: dissimilarities for V39.
  \item V40: dissimilarities for V40.
  \item V41: dissimilarities for V41.
  \item V42: dissimilarities for V42.
  \item V43: dissimilarities for V43.
  \item V44: dissimilarities for V44.
  \item V45: dissimilarities for V45.
}
}
\usage{
food
}
\description{
38 subjects places 45 food items in categories based on similarity.
The dissimilarities are the proportions of combinations NOT placed in the same category.
}
\references{
Ross and Murphy (1999). Food for thought: Cross-classification and category organization in a complex real-world domain. Cognitive psychology, 38(4), 495-553.
Brusco and Stahl (2000). Using Quadratic Assignment Methods to Generate Initial Permutations for Least-Squares Unidimensional Scaling of Symmetric Proximity Matrices. Journal of Classification, 17(2).
}
\keyword{dataset}
