% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fmds.R
\name{print.fmds}
\alias{print.fmds}
\title{Print method for all fmds objects}
\usage{
\method{print}{fmds}(x, ...)
}
\arguments{
\item{x}{object of class \code{fmds}.}

\item{\dots}{additional arguments to be passed.}
}
\value{
none
}
\description{
\code{print} produces output for fmds object.
}
