% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation}
\alias{rotation}
\title{Rotation}
\usage{
rotation(source, weights = NULL, target = NULL, error.check = FALSE)
}
\arguments{
\item{source}{n x m source matrix}

\item{weights}{diagonal of weights matrix, size n}

\item{target}{if NULL: rotate source to principal axes; otherwise: rotate source to n x m target}

\item{error.check}{extensive check validity input parameters (default = FALSE).}
}
\value{
rotation matrix
}
\description{
\code{rotation} returns rotation matrix,
such that || rotation * source - target ||^2_weights is minimal.
}
\references{
Gower (1968).
            Commandeur (1991).
}
\author{
Frank M.T.A. Busing
}
