% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fmds.R
\name{summary.fmds}
\alias{summary.fmds}
\title{Summary method for all fmds objects}
\usage{
\method{summary}{fmds}(object, ...)
}
\arguments{
\item{object}{object of class \code{fmds}.}

\item{\dots}{additional arguments to be passed.}
}
\value{
none
}
\description{
\code{summary} produces an output summary for fmds objects.
}
