% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tortula.R
\docType{data}
\name{tortula}
\alias{tortula}
\title{Tortula data}
\format{
Multivariate data set with 14 cases and 7 variables.
\itemize{
  \item hydroids: (nominal)
  \item leaves: (nominal)
  \item hairs: (ordinal)
  \item apex: (ordinal)
  \item length: (numerical)
  \item diameter: (numerical)
  \item papillae: (numerical)
}
}
\usage{
tortula
}
\description{
Morphological data on 10 taxa of Tortula labeled C, CC, CCS, I, I, N, N, RF, RH, RS, R, R, V, V.
}
\references{
Podani (1999). Extending Gower's general coefficient of similarity to ordinal characters. Taxon, 48(2), 331-340.
}
\keyword{dataset}
