% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petersen.R
\name{petersen}
\alias{petersen}
\title{Mark-Recapture Population Size Estimator}
\usage{
petersen(first, second)
}
\arguments{
\item{first}{vector with first sample identifiers (local minima)}

\item{second}{vector with second sample identifiers (local minima)}
}
\value{
population size estimate
}
\description{
\code{petersen} returns the estimated population size based on two independent equally-sized samples
}
\references{
Busing (2025).
            A Simple Population Size Estimator for Local Minima Applied to Multidimensional Scaling.
}
\author{
Frank M.T.A. Busing
}
